--- python/gui/auto_additions/qgsbasicrelationwidget.py
@@ -1,4 +1,0 @@
-# The following has been generated automatically from src/gui/qgsbasicrelationwidget.h
-QgsBasicRelationWidget.Button.baseClass = QgsBasicRelationWidget
-QgsBasicRelationWidget.Buttons.baseClass = QgsBasicRelationWidget
-Buttons = QgsBasicRelationWidget  # dirty hack since SIP seems to introduce the flags in module
