+++ python/gui/auto_additions/qgsrelationeditorwidget.py
@@ -1,0 +1,4 @@
+# The following has been generated automatically from src/gui/qgsrelationeditorwidget.h
+QgsRelationEditorWidget.Button.baseClass = QgsRelationEditorWidget
+QgsRelationEditorWidget.Buttons.baseClass = QgsRelationEditorWidget
+Buttons = QgsRelationEditorWidget  # dirty hack since SIP seems to introduce the flags in module
