+++ python/gui/gui_auto.sip
@@ -182,7 +182,6 @@
 %Include auto_generated/qgsratiolockbutton.sip
 %Include auto_generated/qgsrelationeditorwidget.sip
 %Include auto_generated/qgsrelationconfigwidget.sip
-%Include auto_generated/qgsbasicrelationwidget.sip
 %Include auto_generated/qgsrelationwidget.sip
 %Include auto_generated/qgsrelationwidgetfactory.sip
 %Include auto_generated/qgsrelationwidgetregistry.sip
