+++ src/gui/qgsrelationeditorwidget.cpp
@@ -1,5 +1,5 @@
 /***************************************************************************
-    qgsrelationeditor.cpp
+    qgsrelationeditorwidget.cpp
      --------------------------------------
     Date                 : 17.5.2013
     Copyright            : (C) 2013 Matthias Kuhn
@@ -58,6 +58,8 @@ QgsFilteredSelectionManager::QgsFilteredSelectionManager( QgsVectorLayer *layer,
 }
 
 const QgsFeatureIds &QgsFilteredSelectionManager::selectedFeatureIds() const
+
+
 {
   return mSelectedFeatureIds;
 }
@@ -91,12 +93,18 @@ void QgsFilteredSelectionManager::onSelectionChanged( const QgsFeatureIds &selec
 
 /// @endcond
 
-QgsRelationEditorWidget::QgsRelationEditorWidget( QWidget *parent )
-  : QgsCollapsibleGroupBox( parent )
+QgsRelationEditorWidget::QgsRelationEditorWidget( const QVariantMap &config, QWidget *parent )
+  : QgsRelationWidget( config, parent )
+  , mButtonsVisibility( qgsFlagKeysToValue( config.value( QStringLiteral( "buttons" ) ).toString(), QgsRelationEditorWidget::Button::AllButtons ) )
 {
-  QVBoxLayout *topLayout = new QVBoxLayout( this );
+  QVBoxLayout *rootLayout = new QVBoxLayout( this );
+  rootLayout->setContentsMargins( 0, 0, 0, 0 );
+
+  mRootCollapsibleGroupBox = new QgsCollapsibleGroupBox( QString(), this );
+  rootLayout->addWidget( mRootCollapsibleGroupBox );
+
+  QVBoxLayout *topLayout = new QVBoxLayout( mRootCollapsibleGroupBox );
   topLayout->setContentsMargins( 0, 9, 0, 0 );
-  setLayout( topLayout );
 
   // buttons
   QHBoxLayout *buttonLayout = new QHBoxLayout();
@@ -198,7 +206,7 @@ QgsRelationEditorWidget::QgsRelationEditorWidget( QWidget *parent )
 
   mRelationLayout->addWidget( mDualView );
 
-  connect( this, &QgsCollapsibleGroupBoxBasic::collapsedStateChanged, this, &QgsRelationEditorWidget::onCollapsedStateChanged );
+  connect( mRootCollapsibleGroupBox, &QgsCollapsibleGroupBoxBasic::collapsedStateChanged, this, &QgsRelationEditorWidget::onCollapsedStateChanged );
   connect( mViewModeButtonGroup, static_cast<void ( QButtonGroup::* )( int )>( &QButtonGroup::buttonClicked ),
            this, static_cast<void ( QgsRelationEditorWidget::* )( int )>( &QgsRelationEditorWidget::setViewMode ) );
   connect( mToggleEditingButton, &QAbstractButton::clicked, this, &QgsRelationEditorWidget::toggleEditing );
@@ -215,48 +223,6 @@ QgsRelationEditorWidget::QgsRelationEditorWidget( QWidget *parent )
 
   // Set initial state for add/remove etc. buttons
   updateButtons();
-}
-
-void QgsRelationEditorWidget::setRelationFeature( const QgsRelation &relation, const QgsFeature &feature )
-{
-  if ( mRelation.isValid() )
-  {
-    disconnect( mRelation.referencingLayer(), &QgsVectorLayer::editingStarted, this, &QgsRelationEditorWidget::updateButtons );
-    disconnect( mRelation.referencingLayer(), &QgsVectorLayer::editingStopped, this, &QgsRelationEditorWidget::updateButtons );
-  }
-
-  mRelation = relation;
-  mFeature = feature;
-
-  connect( mRelation.referencingLayer(), &QgsVectorLayer::editingStarted, this, &QgsRelationEditorWidget::updateButtons );
-  connect( mRelation.referencingLayer(), &QgsVectorLayer::editingStopped, this, &QgsRelationEditorWidget::updateButtons );
-
-  updateTitle();
-
-  QgsVectorLayer *lyr = relation.referencingLayer();
-
-  bool canChangeAttributes = lyr->dataProvider()->capabilities() & QgsVectorDataProvider::ChangeAttributeValues;
-  if ( canChangeAttributes && !lyr->readOnly() )
-  {
-    mToggleEditingButton->setEnabled( true );
-    updateButtons();
-  }
-  else
-  {
-    mToggleEditingButton->setEnabled( false );
-  }
-
-  setObjectName( QStringLiteral( "referenced/" ) + mRelation.name() );
-
-  // If not yet initialized, it is not (yet) visible, so we don't load it to be faster (lazy loading)
-  // If it is already initialized, it has been set visible before and the currently shown feature is changing
-  // and the widget needs updating
-
-  if ( mVisible )
-  {
-    QgsFeatureRequest myRequest = mRelation.getRelatedFeaturesRequest( mFeature );
-    initDualView( mRelation.referencingLayer(), myRequest );
-  }
 }
 
 void QgsRelationEditorWidget::initDualView( QgsVectorLayer *layer, const QgsFeatureRequest &request )
@@ -294,76 +260,6 @@ void QgsRelationEditorWidget::initDualView( QgsVectorLayer *layer, const QgsFeat
   updateButtons();
 }
 
-void QgsRelationEditorWidget::setRelations( const QgsRelation &relation, const QgsRelation &nmrelation )
-{
-  if ( mRelation.isValid() )
-  {
-    disconnect( mRelation.referencingLayer(), &QgsVectorLayer::editingStarted, this, &QgsRelationEditorWidget::updateButtons );
-    disconnect( mRelation.referencingLayer(), &QgsVectorLayer::editingStopped, this, &QgsRelationEditorWidget::updateButtons );
-  }
-
-  if ( mNmRelation.isValid() )
-  {
-    disconnect( mNmRelation.referencedLayer(), &QgsVectorLayer::editingStarted, this, &QgsRelationEditorWidget::updateButtons );
-    disconnect( mNmRelation.referencedLayer(), &QgsVectorLayer::editingStopped, this, &QgsRelationEditorWidget::updateButtons );
-  }
-
-  mRelation = relation;
-  mNmRelation = nmrelation;
-
-  if ( !mRelation.isValid() )
-    return;
-
-  mLayerInSameTransactionGroup = false;
-
-  const auto transactionGroups = QgsProject::instance()->transactionGroups();
-  for ( auto it = transactionGroups.constBegin(); it != transactionGroups.constEnd(); ++it )
-  {
-    if ( mNmRelation.isValid() )
-    {
-      if ( it.value()->layers().contains( mRelation.referencedLayer() ) &&
-           it.value()->layers().contains( mRelation.referencingLayer() ) &&
-           it.value()->layers().contains( mNmRelation.referencedLayer() ) )
-        mLayerInSameTransactionGroup = true;
-    }
-    else
-    {
-      if ( it.value()->layers().contains( mRelation.referencedLayer() ) &&
-           it.value()->layers().contains( mRelation.referencingLayer() ) )
-        mLayerInSameTransactionGroup = true;
-    }
-  }
-
-  connect( mRelation.referencingLayer(), &QgsVectorLayer::editingStarted, this, &QgsRelationEditorWidget::updateButtons );
-  connect( mRelation.referencingLayer(), &QgsVectorLayer::editingStopped, this, &QgsRelationEditorWidget::updateButtons );
-
-  if ( mNmRelation.isValid() )
-  {
-    connect( mNmRelation.referencedLayer(), &QgsVectorLayer::editingStarted, this, &QgsRelationEditorWidget::updateButtons );
-    connect( mNmRelation.referencedLayer(), &QgsVectorLayer::editingStopped, this, &QgsRelationEditorWidget::updateButtons );
-  }
-
-  updateTitle();
-
-  QgsVectorLayer *lyr = relation.referencingLayer();
-
-  bool canChangeAttributes = lyr->dataProvider()->capabilities() & QgsVectorDataProvider::ChangeAttributeValues;
-  if ( canChangeAttributes && !lyr->readOnly() )
-  {
-    mToggleEditingButton->setEnabled( true );
-  }
-  else
-  {
-    mToggleEditingButton->setEnabled( false );
-  }
-
-  updateButtons();
-
-  setObjectName( QStringLiteral( "referenced/" ) + mRelation.name() );
-
-  updateUi();
-}
-
 void QgsRelationEditorWidget::setEditorContext( const QgsAttributeEditorContext &context )
 {
   mEditorContext = context;
@@ -377,30 +273,10 @@ void QgsRelationEditorWidget::setEditorContext( const QgsAttributeEditorContext 
   updateButtons();
 }
 
-QgsAttributeEditorContext QgsRelationEditorWidget::editorContext() const
-{
-  return mEditorContext;
-}
-
-QgsIFeatureSelectionManager *QgsRelationEditorWidget::featureSelectionManager()
-{
-  return mFeatureSelectionMgr;
-}
-
 void QgsRelationEditorWidget::setViewMode( QgsDualView::ViewMode mode )
 {
   mDualView->setView( mode );
   mViewMode = mode;
-}
-
-void QgsRelationEditorWidget::setFeature( const QgsFeature &feature, bool update )
-{
-  mFeature = feature;
-
-  mEditorContext.setFormFeature( feature );
-
-  if ( update )
-    updateUi();
 }
 
 void QgsRelationEditorWidget::updateButtons()
@@ -434,14 +310,15 @@ void QgsRelationEditorWidget::updateButtons()
   mSaveEditsButton->setEnabled( editable );
 
   mToggleEditingButton->setVisible( !mLayerInSameTransactionGroup );
-  mLinkFeatureButton->setVisible( mButtonsVisibility.testFlag( QgsAttributeEditorRelation::Button::Link ) );
-  mUnlinkFeatureButton->setVisible( mButtonsVisibility.testFlag( QgsAttributeEditorRelation::Button::Unlink ) );
-  mSaveEditsButton->setVisible( mButtonsVisibility.testFlag( QgsAttributeEditorRelation::Button::SaveChildEdits ) && !mLayerInSameTransactionGroup );
-  mAddFeatureButton->setVisible( mButtonsVisibility.testFlag( QgsAttributeEditorRelation::Button::AddChildFeature ) );
-  mAddFeatureGeometryButton->setVisible( mButtonsVisibility.testFlag( QgsAttributeEditorRelation::Button::AddChildFeature ) && mEditorContext.mapCanvas() && mEditorContext.cadDockWidget() && spatial );
-  mDuplicateFeatureButton->setVisible( mButtonsVisibility.testFlag( QgsAttributeEditorRelation::Button::DuplicateChildFeature ) );
-  mDeleteFeatureButton->setVisible( mButtonsVisibility.testFlag( QgsAttributeEditorRelation::Button::DeleteChildFeature ) );
-  mZoomToFeatureButton->setVisible( mButtonsVisibility.testFlag( QgsAttributeEditorRelation::Button::ZoomToChildFeature ) && mEditorContext.mapCanvas() && spatial );
+
+  mLinkFeatureButton->setVisible( mButtonsVisibility.testFlag( QgsRelationEditorWidget::Button::Link ) );
+  mUnlinkFeatureButton->setVisible( mButtonsVisibility.testFlag( QgsRelationEditorWidget::Button::Unlink ) );
+  mSaveEditsButton->setVisible( mButtonsVisibility.testFlag( QgsRelationEditorWidget::Button::SaveChildEdits ) && !mLayerInSameTransactionGroup );
+  mAddFeatureButton->setVisible( mButtonsVisibility.testFlag( QgsRelationEditorWidget::Button::AddChildFeature ) );
+  mAddFeatureGeometryButton->setVisible( mButtonsVisibility.testFlag( QgsRelationEditorWidget::Button::AddChildFeature ) && mEditorContext.mapCanvas() && mEditorContext.cadDockWidget() && spatial );
+  mDuplicateFeatureButton->setVisible( mButtonsVisibility.testFlag( QgsRelationEditorWidget::Button::DuplicateChildFeature ) );
+  mDeleteFeatureButton->setVisible( mButtonsVisibility.testFlag( QgsRelationEditorWidget::Button::DeleteChildFeature ) );
+  mZoomToFeatureButton->setVisible( mButtonsVisibility.testFlag( QgsRelationEditorWidget::Button::ZoomToChildFeature ) && mEditorContext.mapCanvas() && spatial );
 }
 
 void QgsRelationEditorWidget::addFeatureGeometry()
@@ -474,60 +351,6 @@ void QgsRelationEditorWidget::addFeatureGeometry()
 
 }
 
-void QgsRelationEditorWidget::addFeature( const QgsGeometry &geometry )
-{
-  QgsAttributeMap keyAttrs;
-
-  const QgsVectorLayerTools *vlTools = mEditorContext.vectorLayerTools();
-
-  if ( mNmRelation.isValid() )
-  {
-    // n:m Relation: first let the user create a new feature on the other table
-    // and autocreate a new linking feature.
-    QgsFeature f;
-    if ( vlTools->addFeature( mNmRelation.referencedLayer(), QgsAttributeMap(), geometry, &f ) )
-    {
-      // Fields of the linking table
-      const QgsFields fields = mRelation.referencingLayer()->fields();
-
-      // Expression context for the linking table
-      QgsExpressionContext context = mRelation.referencingLayer()->createExpressionContext();
-
-      QgsAttributeMap linkAttributes;
-      const auto constFieldPairs = mRelation.fieldPairs();
-      for ( const QgsRelation::FieldPair &fieldPair : constFieldPairs )
-      {
-        int index = fields.indexOf( fieldPair.first );
-        linkAttributes.insert( index,  mFeature.attribute( fieldPair.second ) );
-      }
-
-      const auto constNmFieldPairs = mNmRelation.fieldPairs();
-      for ( const QgsRelation::FieldPair &fieldPair : constNmFieldPairs )
-      {
-        int index = fields.indexOf( fieldPair.first );
-        linkAttributes.insert( index, f.attribute( fieldPair.second ) );
-      }
-      QgsFeature linkFeature = QgsVectorLayerUtils::createFeature( mRelation.referencingLayer(), QgsGeometry(), linkAttributes, &context );
-
-      mRelation.referencingLayer()->addFeature( linkFeature );
-
-      updateUi();
-    }
-  }
-  else
-  {
-    QgsFields fields = mRelation.referencingLayer()->fields();
-
-    const auto constFieldPairs = mRelation.fieldPairs();
-    for ( const QgsRelation::FieldPair &fieldPair : constFieldPairs )
-    {
-      keyAttrs.insert( fields.indexFromName( fieldPair.referencingField() ), mFeature.attribute( fieldPair.referencedField() ) );
-    }
-
-    vlTools->addFeature( mDualView->masterModel()->layer(), keyAttrs, geometry );
-  }
-}
-
 void QgsRelationEditorWidget::onDigitizingCompleted( const QgsFeature &feature )
 {
   addFeature( feature.geometry() );
@@ -535,349 +358,11 @@ void QgsRelationEditorWidget::onDigitizingCompleted( const QgsFeature &feature )
   unsetMapTool();
 }
 
-void QgsRelationEditorWidget::linkFeature()
-{
-  QgsVectorLayer *layer = nullptr;
-
-  if ( mNmRelation.isValid() )
-    layer = mNmRelation.referencedLayer();
-  else
-    layer = mRelation.referencingLayer();
-
-  QgsFeatureSelectionDlg *selectionDlg = new QgsFeatureSelectionDlg( layer, mEditorContext, this );
-  selectionDlg->setAttribute( Qt::WA_DeleteOnClose );
-
-  const QString displayString = QgsVectorLayerUtils::getFeatureDisplayString( mRelation.referencedLayer(), mFeature );
-  selectionDlg->setWindowTitle( tr( "Link existing child features for parent %1 \"%2\"" ).arg( mRelation.referencedLayer()->name(), displayString ) );
-
-  connect( selectionDlg, &QDialog::accepted, this, &QgsRelationEditorWidget::onLinkFeatureDlgAccepted );
-  selectionDlg->show();
-}
-
-void QgsRelationEditorWidget::onLinkFeatureDlgAccepted()
-{
-  QgsFeatureSelectionDlg *selectionDlg = qobject_cast<QgsFeatureSelectionDlg *>( sender() );
-  if ( mNmRelation.isValid() )
-  {
-    QgsFeatureIterator it = mNmRelation.referencedLayer()->getFeatures(
-                              QgsFeatureRequest()
-                              .setFilterFids( selectionDlg->selectedFeatures() )
-                              .setSubsetOfAttributes( mNmRelation.referencedFields() ) );
-
-    QgsFeature relatedFeature;
-
-    QgsFeatureList newFeatures;
-
-    // Fields of the linking table
-    const QgsFields fields = mRelation.referencingLayer()->fields();
-
-    // Expression context for the linking table
-    QgsExpressionContext context = mRelation.referencingLayer()->createExpressionContext();
-
-    QgsAttributeMap linkAttributes;
-    const auto constFieldPairs = mRelation.fieldPairs();
-    for ( const QgsRelation::FieldPair &fieldPair : constFieldPairs )
-    {
-      int index = fields.indexOf( fieldPair.first );
-      linkAttributes.insert( index,  mFeature.attribute( fieldPair.second ) );
-    }
-
-    while ( it.nextFeature( relatedFeature ) )
-    {
-      const auto constFieldPairs = mNmRelation.fieldPairs();
-      for ( const QgsRelation::FieldPair &fieldPair : constFieldPairs )
-      {
-        int index = fields.indexOf( fieldPair.first );
-        linkAttributes.insert( index, relatedFeature.attribute( fieldPair.second ) );
-      }
-      const QgsFeature linkFeature = QgsVectorLayerUtils::createFeature( mRelation.referencingLayer(), QgsGeometry(), linkAttributes, &context );
-
-      newFeatures << linkFeature;
-    }
-
-    mRelation.referencingLayer()->addFeatures( newFeatures );
-    QgsFeatureIds ids;
-    const auto constNewFeatures = newFeatures;
-    for ( const QgsFeature &f : constNewFeatures )
-      ids << f.id();
-    mRelation.referencingLayer()->selectByIds( ids );
-  }
-  else
-  {
-    QMap<int, QVariant> keys;
-    const auto constFieldPairs = mRelation.fieldPairs();
-    for ( const QgsRelation::FieldPair &fieldPair : constFieldPairs )
-    {
-      int idx = mRelation.referencingLayer()->fields().lookupField( fieldPair.referencingField() );
-      QVariant val = mFeature.attribute( fieldPair.referencedField() );
-      keys.insert( idx, val );
-    }
-
-    const auto constSelectedFeatures = selectionDlg->selectedFeatures();
-    for ( QgsFeatureId fid : constSelectedFeatures )
-    {
-      QMapIterator<int, QVariant> it( keys );
-      while ( it.hasNext() )
-      {
-        it.next();
-        mRelation.referencingLayer()->changeAttributeValue( fid, it.key(), it.value() );
-      }
-    }
-  }
-
-  updateUi();
-}
-
-void QgsRelationEditorWidget::duplicateFeature()
-{
-  QgsVectorLayer *layer = mRelation.referencingLayer();
-
-  QgsFeatureIterator fit = layer->getFeatures( QgsFeatureRequest().setFilterFids( mFeatureSelectionMgr->selectedFeatureIds() ) );
-  QgsFeature f;
-  while ( fit.nextFeature( f ) )
-  {
-    QgsVectorLayerUtils::QgsDuplicateFeatureContext duplicatedFeatureContext;
-    QgsVectorLayerUtils::duplicateFeature( layer, f, QgsProject::instance(), duplicatedFeatureContext );
-  }
-}
-
-void QgsRelationEditorWidget::deleteFeature( const QgsFeatureId featureid )
-{
-  deleteFeatures( QgsFeatureIds() << featureid );
-}
-
-void QgsRelationEditorWidget::deleteSelectedFeatures()
-{
-  QgsFeatureIds selectedFids = mFeatureSelectionMgr->selectedFeatureIds();
-  deleteFeatures( selectedFids );
-}
-
-void QgsRelationEditorWidget::deleteFeatures( const QgsFeatureIds &featureids )
-{
-  bool deleteFeatures = true;
-
-  QgsVectorLayer *layer;
-  if ( mNmRelation.isValid() )
-  {
-    layer = mNmRelation.referencedLayer();
-
-    // When deleting a linked feature within an N:M relation,
-    // check if the feature is linked to more than just one feature.
-    // In case it is linked more than just once, ask the user for confirmation
-    // as it is likely he was not aware of the implications and might delete
-    // there may be several linking entries deleted along.
-
-    QgsFeatureRequest deletedFeaturesRequest;
-    deletedFeaturesRequest.setFilterFids( featureids );
-    deletedFeaturesRequest.setFlags( QgsFeatureRequest::NoGeometry );
-    deletedFeaturesRequest.setSubsetOfAttributes( QgsAttributeList() << mNmRelation.referencedFields().first() );
-
-    QgsFeatureIterator deletedFeatures = layer->getFeatures( deletedFeaturesRequest );
-    QStringList deletedFeaturesPks;
-    QgsFeature feature;
-    while ( deletedFeatures.nextFeature( feature ) )
-    {
-      deletedFeaturesPks.append( QgsExpression::quotedValue( feature.attribute( mNmRelation.referencedFields().first() ) ) );
-    }
-
-    QgsFeatureRequest linkingFeaturesRequest;
-    linkingFeaturesRequest.setFlags( QgsFeatureRequest::NoGeometry );
-    linkingFeaturesRequest.setNoAttributes();
-
-    QString linkingFeaturesRequestExpression;
-    if ( !deletedFeaturesPks.empty() )
-    {
-      linkingFeaturesRequestExpression = QStringLiteral( "%1 IN (%2)" ).arg( QgsExpression::quotedColumnRef( mNmRelation.fieldPairs().first().first ), deletedFeaturesPks.join( ',' ) );
-      linkingFeaturesRequest.setFilterExpression( linkingFeaturesRequestExpression );
-
-      QgsFeatureIterator relatedLinkingFeatures = mNmRelation.referencingLayer()->getFeatures( linkingFeaturesRequest );
-
-      int relatedLinkingFeaturesCount = 0;
-      while ( relatedLinkingFeatures.nextFeature( feature ) )
-      {
-        relatedLinkingFeaturesCount++;
-      }
-
-      if ( deletedFeaturesPks.size() == 1 && relatedLinkingFeaturesCount > 1 )
-      {
-        QMessageBox messageBox( QMessageBox::Question, tr( "Really delete entry?" ), tr( "The entry on %1 is still linked to %2 features on %3. Do you want to delete it?" ).arg( mNmRelation.referencedLayer()->name(), QString::number( relatedLinkingFeaturesCount ), mRelation.referencedLayer()->name() ), QMessageBox::NoButton, this );
-        messageBox.addButton( QMessageBox::Cancel );
-        QAbstractButton *deleteButton = messageBox.addButton( tr( "Delete" ),  QMessageBox::AcceptRole );
-
-        messageBox.exec();
-        if ( messageBox.clickedButton() != deleteButton )
-          deleteFeatures = false;
-      }
-      else if ( deletedFeaturesPks.size() > 1 && relatedLinkingFeaturesCount > deletedFeaturesPks.size() )
-      {
-        QMessageBox messageBox( QMessageBox::Question, tr( "Really delete entries?" ), tr( "The %1 entries on %2 are still linked to %3 features on %4. Do you want to delete them?" ).arg( QString::number( deletedFeaturesPks.size() ), mNmRelation.referencedLayer()->name(), QString::number( relatedLinkingFeaturesCount ), mRelation.referencedLayer()->name() ), QMessageBox::NoButton, this );
-        messageBox.addButton( QMessageBox::Cancel );
-        QAbstractButton *deleteButton = messageBox.addButton( tr( "Delete" ), QMessageBox::AcceptRole );
-
-        messageBox.exec();
-        if ( messageBox.clickedButton() != deleteButton )
-          deleteFeatures = false;
-      }
-    }
-  }
-  else
-  {
-    layer = mRelation.referencingLayer();
-  }
-
-  QgsVectorLayerUtils::QgsDuplicateFeatureContext infoContext;
-  if ( QgsVectorLayerUtils::impactsCascadeFeatures( layer, featureids, QgsProject::instance(), infoContext ) )
-  {
-    QString childrenInfo;
-    int childrenCount = 0;
-    const auto infoContextLayers = infoContext.layers();
-    for ( QgsVectorLayer *chl : infoContextLayers )
-    {
-      childrenCount += infoContext.duplicatedFeatures( chl ).size();
-      childrenInfo += ( tr( "%1 feature(s) on layer \"%2\", " ).arg( infoContext.duplicatedFeatures( chl ).size() ).arg( chl->name() ) );
-    }
-
-    // for extra safety to make sure we know that the delete can have impact on children and joins
-    int res = QMessageBox::question( this, tr( "Delete at least %1 feature(s) on other layer(s)" ).arg( childrenCount ),
-                                     tr( "Delete %1 feature(s) on layer \"%2\", %3 as well\nand all of its other descendants.\nDelete these features?" ).arg( featureids.count() ).arg( layer->name() ).arg( childrenInfo ),
-                                     QMessageBox::Yes | QMessageBox::No );
-    if ( res != QMessageBox::Yes )
-      deleteFeatures = false;
-  }
-
-  if ( deleteFeatures )
-  {
-    QgsVectorLayer::DeleteContext context( true, QgsProject::instance() );
-    layer->deleteFeatures( featureids, &context );
-    const auto contextLayers = context.handledLayers();
-    if ( contextLayers.size() > 1 )
-    {
-      int deletedCount = 0;
-      QString feedbackMessage;
-      for ( QgsVectorLayer *contextLayer : contextLayers )
-      {
-        feedbackMessage += tr( "%1 on layer %2. " ).arg( context.handledFeatures( contextLayer ).size() ).arg( contextLayer->name() );
-        deletedCount += context.handledFeatures( contextLayer ).size();
-      }
-      mEditorContext.mainMessageBar()->pushMessage( tr( "%1 features deleted: %2" ).arg( deletedCount ).arg( feedbackMessage ), Qgis::Success );
-    }
-
-    updateUi();
-  }
-}
-
-void QgsRelationEditorWidget::unlinkFeature( const QgsFeatureId featureid )
-{
-  unlinkFeatures( QgsFeatureIds() << featureid );
-}
-
-void QgsRelationEditorWidget::unlinkSelectedFeatures()
-{
-  unlinkFeatures( mFeatureSelectionMgr->selectedFeatureIds() );
-}
-
-void QgsRelationEditorWidget::zoomToSelectedFeatures()
-{
-  QgsMapCanvas *c = mEditorContext.mapCanvas();
-  if ( !c )
-    return;
-
-  c->zoomToFeatureIds(
-    mNmRelation.isValid() ? mNmRelation.referencedLayer() : mRelation.referencingLayer(),
-    mFeatureSelectionMgr->selectedFeatureIds()
-  );
-}
-
-void QgsRelationEditorWidget::unlinkFeatures( const QgsFeatureIds &featureids )
-{
-  if ( mNmRelation.isValid() )
-  {
-    QgsFeatureIterator selectedIterator = mNmRelation.referencedLayer()->getFeatures(
-                                            QgsFeatureRequest()
-                                            .setFilterFids( featureids )
-                                            .setSubsetOfAttributes( mNmRelation.referencedFields() ) );
-
-    QgsFeature f;
-
-    QStringList filters;
-
-    while ( selectedIterator.nextFeature( f ) )
-    {
-      filters << '(' + mNmRelation.getRelatedFeaturesRequest( f ).filterExpression()->expression() + ')';
-    }
-
-    QString filter = QStringLiteral( "(%1) AND (%2)" ).arg(
-                       mRelation.getRelatedFeaturesRequest( mFeature ).filterExpression()->expression(),
-                       filters.join( QLatin1String( " OR " ) ) );
-
-    QgsFeatureIterator linkedIterator = mRelation.referencingLayer()->getFeatures( QgsFeatureRequest()
-                                        .setNoAttributes()
-                                        .setFilterExpression( filter ) );
-
-    QgsFeatureIds fids;
-
-    while ( linkedIterator.nextFeature( f ) )
-    {
-      fids << f.id();
-      QgsDebugMsgLevel( FID_TO_STRING( f.id() ), 4 );
-    }
-
-    mRelation.referencingLayer()->deleteFeatures( fids );
-
-    updateUi();
-  }
-  else
-  {
-    QMap<int, QgsField> keyFields;
-    const auto constFieldPairs = mRelation.fieldPairs();
-    for ( const QgsRelation::FieldPair &fieldPair : constFieldPairs )
-    {
-      int idx = mRelation.referencingLayer()->fields().lookupField( fieldPair.referencingField() );
-      if ( idx < 0 )
-      {
-        QgsDebugMsg( QStringLiteral( "referencing field %1 not found" ).arg( fieldPair.referencingField() ) );
-        return;
-      }
-      QgsField fld = mRelation.referencingLayer()->fields().at( idx );
-      keyFields.insert( idx, fld );
-    }
-
-    const auto constFeatureids = featureids;
-    for ( QgsFeatureId fid : constFeatureids )
-    {
-      QMapIterator<int, QgsField> it( keyFields );
-      while ( it.hasNext() )
-      {
-        it.next();
-        mRelation.referencingLayer()->changeAttributeValue( fid, it.key(), QVariant( it.value().type() ) );
-      }
-    }
-  }
-}
-
 void QgsRelationEditorWidget::toggleEditing( bool state )
 {
-  if ( state )
-  {
-    mEditorContext.vectorLayerTools()->startEditing( mRelation.referencingLayer() );
-    if ( mNmRelation.isValid() )
-      mEditorContext.vectorLayerTools()->startEditing( mNmRelation.referencedLayer() );
-  }
-  else
-  {
-    mEditorContext.vectorLayerTools()->stopEditing( mRelation.referencingLayer() );
-    if ( mNmRelation.isValid() )
-      mEditorContext.vectorLayerTools()->stopEditing( mNmRelation.referencedLayer() );
-  }
+  QgsRelationWidget::toggleEditing( state );
 
   updateButtons();
-}
-
-void QgsRelationEditorWidget::saveEdits()
-{
-  mEditorContext.vectorLayerTools()->saveEdits( mRelation.referencingLayer() );
-  if ( mNmRelation.isValid() )
-    mEditorContext.vectorLayerTools()->saveEdits( mNmRelation.referencedLayer() );
 }
 
 void QgsRelationEditorWidget::onCollapsedStateChanged( bool collapsed )
@@ -926,109 +411,35 @@ void QgsRelationEditorWidget::updateUi()
   }
 }
 
-bool QgsRelationEditorWidget::showLinkButton() const
-{
-  return mLinkFeatureButton->isVisible();
-}
-
-void QgsRelationEditorWidget::setShowLinkButton( bool showLinkButton )
-{
-  mLinkFeatureButton->setVisible( showLinkButton );
-}
-
-bool QgsRelationEditorWidget::showUnlinkButton() const
-{
-  return mUnlinkFeatureButton->isVisible();
-}
-
-void QgsRelationEditorWidget::setShowSaveChildEditsButton( bool showChildEdits )
-{
-  mSaveEditsButton->setVisible( showChildEdits );
-}
-
-bool QgsRelationEditorWidget::showSaveChildEditsButton() const
-{
-  return mSaveEditsButton->isVisible();
-}
-
-void QgsRelationEditorWidget::setVisibleButtons( const QgsAttributeEditorRelation::Buttons &buttons )
+void QgsRelationEditorWidget::setVisibleButtons( const Buttons &buttons )
 {
   mButtonsVisibility = buttons;
   updateButtons();
 }
 
-QgsAttributeEditorRelation::Buttons QgsRelationEditorWidget::visibleButtons() const
+QgsRelationEditorWidget::Buttons QgsRelationEditorWidget::visibleButtons() const
 {
-  QgsAttributeEditorRelation::Buttons buttons;
+  Buttons buttons;
   if ( mLinkFeatureButton->isVisible() )
-    buttons |= QgsAttributeEditorRelation::Button::Link;
+    buttons |= Button::Link;
   if ( mUnlinkFeatureButton->isVisible() )
-    buttons |= QgsAttributeEditorRelation::Button::Unlink;
+    buttons |= Button::Unlink;
   if ( mSaveEditsButton->isVisible() )
-    buttons |= QgsAttributeEditorRelation::Button::SaveChildEdits;
+    buttons |= Button::SaveChildEdits;
   if ( mAddFeatureButton->isVisible() )
-    buttons |= QgsAttributeEditorRelation::Button::AddChildFeature;
+    buttons |= Button::AddChildFeature;
   if ( mDuplicateFeatureButton->isVisible() )
-    buttons |= QgsAttributeEditorRelation::Button::DuplicateChildFeature;
+    buttons |= Button::DuplicateChildFeature;
   if ( mDeleteFeatureButton->isVisible() )
-    buttons |= QgsAttributeEditorRelation::Button::DeleteChildFeature;
+    buttons |= Button::DeleteChildFeature;
   if ( mZoomToFeatureButton->isVisible() )
-    buttons |= QgsAttributeEditorRelation::Button::ZoomToChildFeature;
+    buttons |= Button::ZoomToChildFeature;
   return buttons;
-}
-
-void QgsRelationEditorWidget::setForceSuppressFormPopup( bool forceSuppressFormPopup )
-{
-  mForceSuppressFormPopup = forceSuppressFormPopup;
-}
-
-bool QgsRelationEditorWidget::forceSuppressFormPopup() const
-{
-  return mForceSuppressFormPopup;
-}
-
-void QgsRelationEditorWidget::setNmRelationId( const QVariant &nmRelationId )
-{
-  mNmRelationId = nmRelationId;
-}
-
-QVariant QgsRelationEditorWidget::nmRelationId() const
-{
-  return mNmRelationId;
-}
-
-QString QgsRelationEditorWidget::label() const
-{
-  return mLabel;
-}
-
-void QgsRelationEditorWidget::setLabel( const QString &label )
-{
-  mLabel = label;
-
-  updateTitle();
-}
-
-void QgsRelationEditorWidget::setShowUnlinkButton( bool showUnlinkButton )
-{
-  mUnlinkFeatureButton->setVisible( showUnlinkButton );
 }
 
 void QgsRelationEditorWidget::parentFormValueChanged( const QString &attribute, const QVariant &newValue )
 {
   mDualView->parentFormValueChanged( attribute, newValue );
-}
-
-bool QgsRelationEditorWidget::showLabel() const
-{
-  return mShowLabel;
-}
-
-void QgsRelationEditorWidget::setShowLabel( bool showLabel )
-{
-  mShowLabel = showLabel;
-
-  updateTitle();
 }
 
 void QgsRelationEditorWidget::showContextMenu( QgsActionMenu *menu, const QgsFeatureId fid )
@@ -1067,27 +478,6 @@ void QgsRelationEditorWidget::unsetMapTool()
   disconnect( mMapToolDigitize, &QgsMapToolDigitizeFeature::digitizingCompleted, this, &QgsRelationEditorWidget::onDigitizingCompleted );
 }
 
-void QgsRelationEditorWidget::updateTitle()
-{
-  if ( mShowLabel && !mLabel.isEmpty() )
-  {
-    setTitle( mLabel );
-  }
-  else if ( mShowLabel && mRelation.isValid() )
-  {
-    setTitle( mRelation.name() );
-  }
-  else
-  {
-    setTitle( QString() );
-  }
-}
-
-QgsFeature QgsRelationEditorWidget::feature() const
-{
-  return mFeature;
-}
-
 void QgsRelationEditorWidget::onKeyPressed( QKeyEvent *e )
 {
   if ( e->key() == Qt::Key_Escape )
@@ -1108,3 +498,184 @@ void QgsRelationEditorWidget::mapToolDeactivated()
   }
   mMessageBarItem = nullptr;
 }
+
+QVariantMap QgsRelationEditorWidget::config() const
+{
+  return QVariantMap( {{"buttons", qgsFlagValueToKeys( visibleButtons() )}} );
+}
+
+void QgsRelationEditorWidget::setConfig( const QVariantMap &config )
+{
+  mButtonsVisibility = qgsFlagKeysToValue( config.value( QStringLiteral( "buttons" ) ).toString(), QgsRelationEditorWidget::Button::AllButtons );
+  updateButtons();
+}
+
+void QgsRelationEditorWidget::setTitle( const QString &title )
+{
+  mRootCollapsibleGroupBox->setTitle( title );
+}
+
+void QgsRelationEditorWidget::beforeSetRelationFeature( const QgsRelation &newRelation, const QgsFeature &newFeature )
+{
+  Q_UNUSED( newRelation );
+  Q_UNUSED( newFeature );
+
+  if ( ! mRelation.isValid() )
+    return;
+
+  disconnect( mRelation.referencingLayer(), &QgsVectorLayer::editingStarted, this, &QgsRelationEditorWidget::updateButtons );
+  disconnect( mRelation.referencingLayer(), &QgsVectorLayer::editingStopped, this, &QgsRelationEditorWidget::updateButtons );
+}
+
+void QgsRelationEditorWidget::afterSetRelationFeature()
+{
+  mToggleEditingButton->setEnabled( false );
+
+  if ( ! mRelation.isValid() )
+    return;
+
+  connect( mRelation.referencingLayer(), &QgsVectorLayer::editingStarted, this, &QgsRelationEditorWidget::updateButtons );
+  connect( mRelation.referencingLayer(), &QgsVectorLayer::editingStopped, this, &QgsRelationEditorWidget::updateButtons );
+
+  QgsVectorLayer *vl = mRelation.referencingLayer();
+  bool canChangeAttributes = vl->dataProvider()->capabilities() & QgsVectorDataProvider::ChangeAttributeValues;
+  if ( canChangeAttributes && !vl->readOnly() )
+  {
+    mToggleEditingButton->setEnabled( true );
+    updateButtons();
+  }
+  else
+  {
+    mToggleEditingButton->setEnabled( false );
+  }
+
+  // If not yet initialized, it is not (yet) visible, so we don't load it to be faster (lazy loading)
+  // If it is already initialized, it has been set visible before and the currently shown feature is changing
+  // and the widget needs updating
+
+  if ( mVisible )
+  {
+    QgsFeatureRequest myRequest = mRelation.getRelatedFeaturesRequest( mFeature );
+    initDualView( mRelation.referencingLayer(), myRequest );
+  }
+}
+
+void QgsRelationEditorWidget::beforeSetRelations( const QgsRelation &newRelation, const QgsRelation &newNmRelation )
+{
+  Q_UNUSED( newRelation );
+  Q_UNUSED( newNmRelation );
+
+  if ( mRelation.isValid() )
+  {
+    disconnect( mRelation.referencingLayer(), &QgsVectorLayer::editingStarted, this, &QgsRelationEditorWidget::updateButtons );
+    disconnect( mRelation.referencingLayer(), &QgsVectorLayer::editingStopped, this, &QgsRelationEditorWidget::updateButtons );
+  }
+
+  if ( mNmRelation.isValid() )
+  {
+    disconnect( mNmRelation.referencedLayer(), &QgsVectorLayer::editingStarted, this, &QgsRelationEditorWidget::updateButtons );
+    disconnect( mNmRelation.referencedLayer(), &QgsVectorLayer::editingStopped, this, &QgsRelationEditorWidget::updateButtons );
+  }
+}
+
+void QgsRelationEditorWidget::afterSetRelations()
+{
+  if ( !mRelation.isValid() )
+    return;
+
+  connect( mRelation.referencingLayer(), &QgsVectorLayer::editingStarted, this, &QgsRelationEditorWidget::updateButtons );
+  connect( mRelation.referencingLayer(), &QgsVectorLayer::editingStopped, this, &QgsRelationEditorWidget::updateButtons );
+
+  if ( mNmRelation.isValid() )
+  {
+    connect( mNmRelation.referencedLayer(), &QgsVectorLayer::editingStarted, this, &QgsRelationEditorWidget::updateButtons );
+    connect( mNmRelation.referencedLayer(), &QgsVectorLayer::editingStopped, this, &QgsRelationEditorWidget::updateButtons );
+  }
+
+  QgsVectorLayer *vl = mRelation.referencingLayer();
+  bool canChangeAttributes = vl->dataProvider()->capabilities() & QgsVectorDataProvider::ChangeAttributeValues;
+  if ( canChangeAttributes && !vl->readOnly() )
+  {
+    mToggleEditingButton->setEnabled( true );
+  }
+  else
+  {
+    mToggleEditingButton->setEnabled( false );
+  }
+
+  updateButtons();
+}
+
+
+///////////////////////////////////////////////////////////////////////////////
+
+
+#ifndef SIP_RUN
+QgsRelationEditorWidgetFactory::QgsRelationEditorWidgetFactory()
+{
+
+}
+
+QString QgsRelationEditorWidgetFactory::type() const
+{
+  return QStringLiteral( "relation_editor" );
+}
+
+QString QgsRelationEditorWidgetFactory::name() const
+{
+  return QStringLiteral( "Relation Editor" );
+}
+
+QgsRelationWidget *QgsRelationEditorWidgetFactory::create( const QVariantMap &config, QWidget *parent ) const
+{
+  return new QgsRelationEditorWidget( config, parent );
+}
+
+
+QgsRelationConfigWidget *QgsRelationEditorWidgetFactory::configWidget( const QgsRelation &relation, QWidget *parent ) const
+{
+  return static_cast<QgsRelationConfigWidget *>( new QgsRelationEditorConfigWidget( relation, parent ) );
+}
+#endif
+
+
+///////////////////////////////////////////////////////////////////////////////
+
+
+#ifndef SIP_RUN
+QgsRelationEditorConfigWidget::QgsRelationEditorConfigWidget( const QgsRelation &relation, QWidget *parent )
+  : QgsRelationConfigWidget( relation, parent )
+{
+  setupUi( this );
+}
+
+QVariantMap QgsRelationEditorConfigWidget::config()
+{
+  QgsRelationEditorWidget::Buttons buttons;
+  buttons.setFlag( QgsRelationEditorWidget::Button::Link, mRelationShowLinkCheckBox->isChecked() );
+  buttons.setFlag( QgsRelationEditorWidget::Button::Unlink, mRelationShowUnlinkCheckBox->isChecked() );
+  buttons.setFlag( QgsRelationEditorWidget::Button::AddChildFeature, mRelationShowAddChildCheckBox->isChecked() );
+  buttons.setFlag( QgsRelationEditorWidget::Button::DuplicateChildFeature, mRelationShowDuplicateChildFeatureCheckBox->isChecked() );
+  buttons.setFlag( QgsRelationEditorWidget::Button::ZoomToChildFeature, mRelationShowZoomToFeatureCheckBox->isChecked() );
+  buttons.setFlag( QgsRelationEditorWidget::Button::DeleteChildFeature, mRelationDeleteChildFeatureCheckBox->isChecked() );
+  buttons.setFlag( QgsRelationEditorWidget::Button::SaveChildEdits, mRelationShowSaveChildEditsCheckBox->isChecked() );
+
+  return QVariantMap(
+  {
+    {"buttons", qgsFlagValueToKeys( buttons )},
+  } );
+}
+
+void QgsRelationEditorConfigWidget::setConfig( const QVariantMap &config )
+{
+  const QgsRelationEditorWidget::Buttons buttons = qgsFlagKeysToValue( config.value( QStringLiteral( "buttons" ) ).toString(), QgsRelationEditorWidget::Button::AllButtons );
+
+  mRelationShowLinkCheckBox->setChecked( buttons.testFlag( QgsRelationEditorWidget::Button::Link ) );
+  mRelationShowUnlinkCheckBox->setChecked( buttons.testFlag( QgsRelationEditorWidget::Button::Unlink ) );
+  mRelationShowAddChildCheckBox->setChecked( buttons.testFlag( QgsRelationEditorWidget::Button::AddChildFeature ) );
+  mRelationShowDuplicateChildFeatureCheckBox->setChecked( buttons.testFlag( QgsRelationEditorWidget::Button::DuplicateChildFeature ) );
+  mRelationShowZoomToFeatureCheckBox->setChecked( buttons.testFlag( QgsRelationEditorWidget::Button::ZoomToChildFeature ) );
+  mRelationDeleteChildFeatureCheckBox->setChecked( buttons.testFlag( QgsRelationEditorWidget::Button::DeleteChildFeature ) );
+  mRelationShowSaveChildEditsCheckBox->setChecked( buttons.testFlag( QgsRelationEditorWidget::Button::SaveChildEdits ) );
+}
+#endif
